[connector] Move everything up a level
